finalize this project: it should be a skeleton project as a JSWEET plus angular quick start.
A java-to-ts-or-js transpiler may make sense in future, but for now it
take too much time and got only little benefit.

Type script is good enough for code edit and maintain thanks to IDE
support, re-factoring is code navigation is quite good.

Angular is totally based on ts hence making it hard to translate so many
features into java code manually.